Refactor: TextInput, constants 파일 분리
